refactor(chat): split research answer helpers
