+++ server/chat/report-research-answer.ts
@@ -1,342 +1,28 @@
 import { extractWeatherLocation } from "../../lib/tools/realtime-info.js";
-
-type ResearchAnswerKind =
-  | ""
-  | "stock"
-  | "real_estate"
-  | "market_weather_brief"
-  | "weather"
-  | "sports"
-  | "market"
-  | "news"
-  | "generic"
-  | (string & {});
-
-interface ToolTextContent {
-  text?: unknown;
-  [key: string]: unknown;
-}
-
-interface DirectQuoteSnapshot {
-  symbol?: unknown;
-  close?: unknown;
-  date?: unknown;
-  time?: unknown;
-  source?: unknown;
-  url?: unknown;
-  open?: unknown;
-  high?: unknown;
-  low?: unknown;
-  [key: string]: unknown;
-}
-
-interface EvidenceSource {
-  title?: unknown;
-  source?: unknown;
-  url?: unknown;
-  [key: string]: unknown;
-}
-
-interface ToolResultDetails {
-  directQuotes?: DirectQuoteSnapshot[];
-  provider?: unknown;
-  sources?: EvidenceSource[];
-  location?: unknown;
-  [key: string]: unknown;
-}
-
-interface ToolExecutionResult {
-  content?: ToolTextContent[];
-  details?: ToolResultDetails;
-  [key: string]: unknown;
-}
-
-interface StooqItem {
-  symbol: string;
-  source: string;
-  url: string;
-  name: string;
-  price: string;
-  change: string;
-  timestamp: string;
-  range: string;
-}
-
-interface StockSnapshot {
-  symbol?: unknown;
-  price?: unknown;
-  timestamp?: string;
-  source?: unknown;
-  url?: unknown;
-  range?: string;
-}
-
-interface IndexFallbackTarget {
-  label?: string;
-  [key: string]: unknown;
-}
-
-interface IndexSnapshot {
-  name?: string;
-  level?: string;
-  change?: string;
-  source?: unknown;
-  url?: unknown;
-  queryDate: string;
-}
-
-interface WeatherForecastRow {
-  date: string;
-  desc: string;
-  min: string;
-  max: string;
-}
-
-interface WeatherSnapshot {
-  location?: unknown;
-  date: string;
-  desc: string;
-  tempRange: string;
-}
-
-interface EvidenceMeta {
-  tool: string;
-  basis: string;
-  caveat: string;
-}
-
-interface EvidenceBlockOptions {
-  kind?: ResearchAnswerKind;
-  context?: unknown;
-  userPrompt?: unknown;
-}
-
-type StructuredSectionEntry = readonly [string, unknown];
-type StructuredFields = Record<string, string>;
-
-interface TempRange {
-  min: number | null;
-  max: number | null;
-}
-
-interface GoldSummary {
-  date: string;
-  jewelry: string;
-  bars: string;
-  recovery: string;
-  examples: string;
-  sge: string;
-  shuibei: string;
-  international: string;
-}
-
-interface NewsItem {
-  title: string;
-  source: string;
-  sourceUrl: string;
-  link: string;
-  snippet?: string;
-  published: string;
-  windowLabel: string;
-  freshness: string;
-}
-
-interface BuildAnswerOptions {
-  userPrompt?: unknown;
-  prompt?: unknown;
-  [key: string]: unknown;
-}
-
-const MARKET_WEATHER_TICKER_STOPWORDS = new Set([
-  "AI", "API", "ETF", "ETFS", "USD", "CNY", "EUR", "GBP", "JPY",
-  "PE", "PB", "PS", "IPO", "CEO", "CFO", "GDP", "CPI", "PPI",
-  "MACD", "RSI", "UTC", "PPT", "TOOL",
-]);
-const COMMON_US_TICKERS = new Set([
-  "AAPL", "TSLA", "NVDA", "MSFT", "GOOGL", "GOOG", "AMZN", "META",
-  "NFLX", "AMD", "INTC", "AVGO", "SMCI", "PLTR", "COIN", "MSTR",
-  "BABA", "PDD", "NIO", "XPEV", "LI",
-]);
-function textOf(value: unknown): string {
-  return String(value || "").replace(/\s+/g, " ").trim();
-}
-function extractToolText(result: ToolExecutionResult | null | undefined): string {
-  return (result?.content || [])
-    .map((item) => item?.text || "")
-    .filter(Boolean)
-    .join("\n")
-    .trim();
-}
-function parseStooqItems(context: unknown): StooqItem[] {
-  const text = String(context || "");
-  const starts = Array.from(text.matchAll(/(?:^|\n)(\d+\.\s+[A-Z0-9.]{1,12}\s+最近可用行情)/g)).map((match) => {
-    const full = String(match[0] || "");
-    const header = String(match[1] || "");
-    const rawIndex = Number(match.index || 0);
-    return full.startsWith("\n") ? rawIndex + 1 : rawIndex + full.indexOf(header);
-  });
-  const blocks = starts.map((start, index) => {
-    const end = index + 1 < starts.length ? starts[index + 1] : text.length;
-    return text.slice(start, end).trim();
-  }).filter(Boolean);
-  return blocks.map((block) => {
-    const symbol = block.match(/^\d+\.\s+([A-Z0-9.]{1,12})\s+最近可用行情/m)?.[1] || "";
-    const source = block.match(/^来源[:：]\s*([^\n]+)/m)?.[1] || "";
-    const url = block.match(/^(https?:\/\/\S+)/m)?.[1] || "";
-    const name = block.match(/^-+\s*名称[:：]\s*([^\n]+)/m)?.[1] || "";
-    const price = block.match(/^-+\s*价格[:：]\s*([^\n]+)/m)?.[1] || "";
-    const change = block.match(/^-+\s*涨跌\/涨跌幅[:：]\s*([^\n]+)/m)?.[1] || "";
-    const timestamp = block.match(/^-+\s*时间戳[:：]\s*([^\n]+)/m)?.[1] || "";
-    const range = block.match(/^-+\s*开盘\/最高\/最低[:：]\s*([^\n]+)/m)?.[1] || "";
-    return { symbol, source, url, name, price, change, timestamp, range };
-  }).filter((item) => item.symbol && item.price);
-}
-function extractRequestedUsTickers(text: unknown): string[] {
-  const symbols: string[] = [];
-  for (const match of String(text || "").matchAll(/\$?\b([A-Z]{1,5})(?:\.US)?\b/g)) {
-    const raw = String(match[0] || "");
-    const bare = String(match[1] || "").toUpperCase();
-    if (MARKET_WEATHER_TICKER_STOPWORDS.has(bare)) continue;
-    if (!raw.startsWith("$") && !COMMON_US_TICKERS.has(bare)) continue;
-    if (!symbols.includes(bare)) symbols.push(bare);
-  }
-  return symbols.slice(0, 8);
-}
-function escapeRegExp(value: unknown): string {
-  return String(value || "").replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-function extractStructuredContextSection(context: unknown, title: string): string {
-  const re = new RegExp(`【${escapeRegExp(title)}】\\n([\\s\\S]*?)(?=\\n【|$)`);
-  return String(context || "").match(re)?.[1]?.trim() || "";
-}
-function parseStructuredFields(sectionText: unknown): StructuredFields {
-  const fields: StructuredFields = {};
-  for (const rawLine of String(sectionText || "").split(/\r?\n/)) {
-    const line = textOf(rawLine);
-    const match = line.match(/^-?\s*([^:：]+)[:：]\s*(.+)$/);
-    if (match) fields[match[1].trim()] = match[2].trim();
-  }
-  return fields;
-}
-function buildStructuredSection(title: string, entries: readonly StructuredSectionEntry[]): string {
-  const lines = entries
-    .filter(([, value]) => textOf(value))
-    .map(([label, value]) => `- ${label}: ${textOf(value)}`);
-  if (!lines.length) return "";
-  return [`【${title}】`, ...lines].join("\n");
-}
-function formatLocalDate(date: Date = new Date()): string {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  return `${year}-${month}-${day}`;
-}
-function formatLocalDateTime(date: Date = new Date()): string {
-  const hour = String(date.getHours()).padStart(2, "0");
-  const minute = String(date.getMinutes()).padStart(2, "0");
-  return `${formatLocalDate(date)} ${hour}:${minute}`;
-}
-function uniqueCompact(items: unknown[], max: number = 3): string[] {
-  const seen = new Set<string>();
-  const out: string[] = [];
-  for (const item of items.map(textOf).filter(Boolean)) {
-    const clean = item
-      .replace(/\s+/g, " ")
-      .replace(/^\d+\.\s*/, "")
-      .replace(/[，。；;,.]+$/g, "")
-      .trim();
-    if (!clean || /^(?:com|cn|net|org|www)$/i.test(clean) || seen.has(clean)) continue;
-    seen.add(clean);
-    out.push(clean);
-    if (out.length >= max) break;
-  }
-  return out;
-}
-function extractEvidenceSources(context: unknown, max: number = 3): string[] {
-  const text = String(context || "");
-  const sources: unknown[] = [];
-  for (const match of text.matchAll(/(?:来源|来源站点)[:：]\s*([^\n]+)/g)) {
-    sources.push(match[1]);
-  }
-  for (const match of text.matchAll(/\[(Stooq|新浪财经|东方财富|上海黄金交易所|上金所|gold-api|Open-Meteo|wttr\.in|中国天气|中央气象台)[^\]]*\]/gi)) {
-    sources.push(match[1]);
-  }
-  for (const match of text.matchAll(/\b(https?:\/\/[^\s)）]+)\b/g)) {
-    try {
-      sources.push(new URL(match[1]).hostname.replace(/^www\./, ""));
-    } catch {
-      // Ignore malformed snippets from search output.
-    }
-  }
-  return uniqueCompact(sources, max);
-}
-function inferEvidenceMeta(kind: ResearchAnswerKind | undefined, prompt: string): EvidenceMeta {
-  if (kind === "weather") {
-    return {
-      tool: "weather",
-      basis: "天气工具返回的预报快照，按用户问法优先选择今天/明天/后天对应日期。",
-      caveat: "天气预报会滚动变化，出门前建议再看本地天气 App 或雷达。",
-    };
-  }
-  if (kind === "news") {
-    return {
-      tool: "live_news / web_search",
-      basis: "按最近新闻候选、发布时间、来源链接和主题相关度筛选。",
-      caveat: "正式引用前建议打开原文核验全文和发布时间。",
-    };
-  }
-  if (kind === "market_weather_brief") {
-    return {
-      tool: "stock_market + weather",
-      basis: "综合行情快照和天气预报快照生成行动建议。",
-      caveat: "行情与天气都会实时波动，关键决策前请二次核验。",
-    };
-  }
-  if (/金价|黄金|白银|金交所|金店|回收价|Au99\.99|Au9999|XAU|金条/i.test(prompt)) {
-    return {
-      tool: "stock_market",
-      basis: "黄金/白银行情与品牌报价聚合结果，优先使用含明确数字的最近可用报价。",
-      caveat: "门店、地区、工费和更新时间会造成差异，不构成投资或购买建议。",
-    };
-  }
-  if (/原油|油价|布伦特|WTI|crude|oil/i.test(prompt)) {
-    return {
-      tool: "stock_market",
-      basis: "原油行情工具返回的最近可用合约报价。",
-      caveat: "期货/CFD 价格盘中波动明显，交易前请用行情终端核验。",
-    };
-  }
-  if (kind === "market") {
-    return {
-      tool: "stock_market",
-      basis: "行情工具返回的最近可用股价、指数或板块候选数据。",
-      caveat: "行情展示不构成投资建议；交易级实时性请用券商或交易所源核验。",
-    };
-  }
-  return {
-    tool: "research_prefetch",
-    basis: "系统预取资料中的可用证据。",
-    caveat: "资料不足时应继续补充来源再下结论。",
-  };
-}
-function appendEvidenceBlock(answer: unknown, { kind, context, userPrompt }: EvidenceBlockOptions = {}): string {
-  const body = String(answer || "").trim();
-  if (!body || /数据来源\/判断依据/.test(body)) return body;
-  const prompt = textOf(userPrompt);
-  if (/(?:一句话|一行|简短|简洁|直接回答|只回复|只回答)/.test(prompt)) return body;
-  const meta = inferEvidenceMeta(kind, prompt);
-  const sources = extractEvidenceSources(context);
-  return [
-    body,
-    "",
-    "数据来源/判断依据",
-    `- 工具：${meta.tool}`,
-    `- 时间：${formatLocalDateTime()}（本机时间）`,
-    sources.length ? `- 来源：${sources.join("、")}` : "",
-    `- 依据：${meta.basis}`,
-    `- 注意：${meta.caveat}`,
-  ].filter(Boolean).join("\n");
-}
+import { appendEvidenceBlock } from "./report-research-evidence.js";
+import type {
+  BuildAnswerOptions,
+  GoldSummary,
+  IndexFallbackTarget,
+  IndexSnapshot,
+  NewsItem,
+  ResearchAnswerKind,
+  StockSnapshot,
+  TempRange,
+  ToolExecutionResult,
+  WeatherForecastRow,
+  WeatherSnapshot,
+} from "./report-research-answer-types.js";
+import {
+  buildStructuredSection,
+  extractRequestedUsTickers,
+  extractStructuredContextSection,
+  extractToolText,
+  formatLocalDate,
+  parseStooqItems,
+  parseStructuredFields,
+  textOf,
+} from "./report-research-answer-utils.js";
 function parseStockSnapshot(result: ToolExecutionResult | null | undefined): StockSnapshot | null {
   const directQuote = result?.details?.directQuotes?.[0];
   if (directQuote?.symbol && directQuote?.close) {
